Move mentions files to client/server
