+++ packages/rocketchat-mentions/server/Mentions.js
